backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-dotenv
PyJWT
passlib[bcrypt]
python-multipart
